// src/modules/admin/scraped-data/dto/create-unregistered.dto.ts
// ======================================================
// DTO
// src/modules/unregistered/dto/create-unregistered.dto.ts
// ======================================================

export class CreateUnregisteredDto {

  country?: string;

  state?: string;

  district?: string;

  subDistrict?: string;

  eventType?: string;

}